refactor: method 모듈 auth로 변경
