+++ src/main/java/com/github/bluekey/service/auth/AuthService.java
@@ -1,12 +1,16 @@
-package com.github.bluekey.service.member;
+package com.github.bluekey.service.auth;
 
 import com.github.bluekey.dto.JwtInfoDto;
 import com.github.bluekey.dto.LoginMemberDto;
 import com.github.bluekey.dto.request.LoginRequestDto;
+import com.github.bluekey.dto.request.SignupRequestDto;
 import com.github.bluekey.dto.response.LoginTokenResponseDto;
 import com.github.bluekey.dto.response.LoginTokenResponseDto;
+import com.github.bluekey.dto.response.SignupResponseDto;
 import com.github.bluekey.entity.member.Member;
+import com.github.bluekey.entity.member.MemberType;
 import com.github.bluekey.exception.AuthenticationException;
+import com.github.bluekey.exception.BusinessException;
 import com.github.bluekey.exception.ErrorCode;
 import com.github.bluekey.jwt.JwtProvider;
 import com.github.bluekey.repository.member.MemberRepository;
@@ -14,6 +18,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
@@ -32,6 +37,49 @@ public class AuthService {
 		return generateLoginTokenResponseDto(member, token);
 	}
 
+	@Transactional
+	public SignupResponseDto createAdmin(SignupRequestDto dto) {
+		validateSignUpRequest(dto);
+		Member admin = dto.toMember();
+		encodeMemberPassword(admin);
+		Member newMember = memberRepository.save(admin);
+		return SignupResponseDto.from(newMember);
+	}
+
+	public String getEncodePassword(String password) {
+		return passwordEncoder.encode(password);
+	}
+
+	private void encodeMemberPassword(Member member) {
+		member.updatePassword(getEncodePassword(member.getPassword()));
+	}
+
+	private void validateSignUpRequest(SignupRequestDto dto) {
+		validateAdminLoginId(dto.getLoginId());
+		validateAdminEmail(dto.getEmail());
+		validateAdminNickname(dto.getNickname());
+	}
+
+	private void validateAdminLoginId(String loginId) {
+		memberRepository.findMemberByLoginId(loginId)
+				.ifPresent(member -> {throw new BusinessException(ErrorCode.INVALID_LOGIN_ID_VALUE);
+				});
+	}
+
+	private void validateAdminEmail(String email) {
+		memberRepository.findMemberByEmailAndType(email, MemberType.ADMIN)
+				.ifPresent(member -> {throw new BusinessException(ErrorCode.INVALID_EMAIL_VALUE);
+				});
+	}
+
+	private void validateAdminNickname(String nickname) {
+		// 아티스트의 활동 예명을 닉네임으로 사용할 수 없다.
+		if (memberRepository.findMemberByNameAndType(nickname, MemberType.USER).isPresent() ||
+				memberRepository.findMemberByEnNameAndType(nickname, MemberType.USER).isPresent()) {
+			throw new BusinessException(ErrorCode.INVALID_NICKNAME_VALUE);
+		}
+	}
+
 	private Member validateLogin(LoginRequestDto dto){
 		Member member = memberRepository.findMemberByLoginId(dto.getLoginId())
 				.orElseThrow(() -> new AuthenticationException(ErrorCode.AUTHENTICATION_FAILED, AUTHENTICATION_ERROR_MESSAGE));
@@ -47,4 +95,6 @@ public class AuthService {
 				.jwtInformation(JwtInfoDto.builder().accessToken(token).build())
 				.build();
 	}
+
+
 }
